Adds option for the exposed ingress routes
The Legend Studio application typically serves on `/studio`, however
it can serve `/` as well, so a 404 Error is not encountered while
accessing the root URI.

However, for the Public instance we would like to route the `/` route
to the github pages, which means that we need to be able to tune what
the Legend Studio charm exposes as an ingress route.

Adds use-root-path config option (True by default). Setting it to
False will result in the charm exposing the `/studio` ingress route
instead of `/`.